Display yapf formatting error class and message
This PR fixes an error where if yapf raised an exception it would result in
a unbound variable error for the variable 'error', because it had the same name
as the Exception aliased as 'error'. This PR uses different names for the Exception
alias and the error string. It also adds the error class as template context to
display the error type to the User.